fix song reconiser

+++ user_interface.py
@@ -13,56 +13,71 @@ import numpy as np
 import settings
 from database_storage import AudioDatabase
 
-#test
 # Eine Überschrift der ersten Ebene
 st.write("Musikerkennung")
 
 # tabs
 tab1, tab2 = st.tabs(["Hochladen", "Erkennen"])
 
-
 with tab1:
     st.write("Hochladen")
-    
-      
-    
-    with st.form ("upload"):
 
-        uploaded_song = st.file_uploader("Choose a file", type=['mp3', 'wav'])
-        
+    with st.form("upload"):
+        uploaded_song = st.file_uploader("Choose a file", type=['wav'])
+
         if uploaded_song is not None:
             file_path = os.path.join(os.getcwd(), uploaded_song.name)
-            
-            with open(uploaded_song.name, 'wb') as f:
+
+            with open(file_path, 'wb') as f:
                 f.write(uploaded_song.getvalue())
-            
+
             st.write("file ausgewählt")
-            
-        
-        title = st.text_input("# Title") 
-        
-        interpret = st.text_input("# Interpret")   
-   
+
+            title = st.text_input("# Title")
+            interpret = st.text_input("# Interpret")
+
         submitted = st.form_submit_button("Submit")
-        
+
         if submitted:
-                
-                song_info= ",".join([title,interpret])
-                print (uploaded_song)                     
-                fingerprinter_instance = AudioFingerprinter()
-                fingerprint = fingerprinter_instance.fingerprint_file(file_path)
-                new_song = AudioDatabase("shazam/database.json")
-                new_song.store_song(fingerprint,song_info)
-                
-                print (fingerprint)
-                #new_song= Data(title,interpret,fingerprint)
-                #new_song.store_data()         
-                st.write("upload complete")    
-                os.remove(file_path)
-                st.rerun()
+            song_info = f"{title},{interpret}"
+            fingerprinter_instance = AudioFingerprinter()
+            fingerprint = fingerprinter_instance.fingerprint_file(file_path)
+            new_song = AudioDatabase("database.json")
+            new_song.store_song(fingerprint, song_info)
+            st.write("upload complete")
+            os.remove(file_path)
+            st.experimental_rerun()
 
+with tab2:
+    st.write("Erkennen")
 
-    #Brauch noch exeption handling
+    with st.form("recognize_song"):
+        uploaded_song = st.file_uploader("Wählen Sie eine Datei aus")
+
+        if uploaded_song is not None:
+            file_path = os.path.join(os.getcwd(), uploaded_song.name)
+
+            with open(file_path, 'wb') as f:
+                
+                    f.write(uploaded_song.getvalue())
+
+            st.write("Datei ausgewählt:", uploaded_song.name)
+
+        submitted = st.form_submit_button("Song erkennen")
+
+        if submitted and uploaded_song:
+            recognizer = SongRecognizer("database.json")
+            recognition_result = recognizer.recognise_song(file_path)
+            os.remove(file_path)
+
+            if recognition_result is not None:
+                st.write("Der hochgeladene Song wurde erkannt!")
+                st.write("Künstler:", recognition_result[0])
+                st.write("Album:", recognition_result[1])
+                st.write("Titel:", recognition_result[2])
+            else:
+                st.write("Der hochgeladene Song wurde nicht in der Datenbank gefunden.")
+
 """      
     if title and interpret is not None:
             with DDGS() as ddgs:
@@ -95,7 +110,7 @@ with tab1:
                         st.write("Album Cover URL nicht gefunden.")
                 else:
                     st.write("Album Cover nicht gefunden.")
-"""
+
 
  
 
@@ -129,3 +144,4 @@ with tab2:
                 st.write("Titel:", recognition_result[2])
             else:
                 st.write("Der hochgeladene Song wurde nicht in der Datenbank gefunden.")
+                """
